services/gemini_service.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()
GENAI_API_KEY = os.getenv("GEMINI_API_KEY")
genai.configure(api_key=GENAI_API_KEY)
model = genai.GenerativeModel(model_name="gemini-1.5-flash")


def generate_answer(prompt):
    response = model.generate_content(prompt)
    return response.text
